changed record type for faster check and moved test cases, added test for pipeline enforce
