+++ forte/data/data_pack.py
@@ -63,7 +63,7 @@ class Meta(BaseMeta):
         super().__init__(pack_name)
         self.language = language
         self.span_unit = span_unit
-        self.record: Dict[str, List[str]] = dict()
+        self.record: Dict[str, Set[str]] = dict()
 
 
 class DataPack(BasePack[Entry, Link, Group]):
