+++ forte/evaluation/base/base_evaluator.py
@@ -15,7 +15,7 @@
 Defines the Evaluator interface and related functions.
 """
 from abc import abstractmethod
-from typing import Any, Dict, List
+from typing import Any, Dict, Set
 
 from forte.data.base_pack import PackType
 from forte.pipeline_component import PipelineComponent
@@ -30,8 +30,8 @@ class Evaluator(PipelineComponent[PackType]):
     r"""The base class of the evaluator."""
     def __init__(self):
         super().__init__()
-        self._pred_pack_expectation: Dict[str, List[str]] = None
-        self._ref_pack_expectation: Dict[str, List[str]] = None
+        self._pred_pack_expectation: Dict[str, Set[str]] = None
+        self._ref_pack_expectation: Dict[str, Set[str]] = None
 
     @abstractmethod
     def consume_next(self, pred_pack: PackType, ref_pack: PackType):
@@ -57,9 +57,9 @@ class Evaluator(PipelineComponent[PackType]):
 
     def expected_types_and_attributes(self,
                                       pred_pack_expectation: Dict[
-                                          str, List[str]],
+                                          str, Set[str]],
                                       ref_pack_expectation: Dict[
-                                          str, List[str]]):
+                                          str, Set[str]]):
         r"""If the evaluator has required input types and attributes for
         `pred_pack` or `ref_pack`, user could add the types and attributes
         required with this function.
@@ -73,7 +73,7 @@ class Evaluator(PipelineComponent[PackType]):
         self._pred_pack_expectation = pred_pack_expectation
         self._ref_pack_expectation = ref_pack_expectation
 
-    def pred_pack_record(self, record_meta: Dict[str, List[str]]):
+    def pred_pack_record(self, record_meta: Dict[str, Set[str]]):
         r"""Method to add output type record of prediction datapack of
         current processor to :attr:`forte.data.base_pack.BaseMeta.record`.
 
@@ -83,7 +83,7 @@ class Evaluator(PipelineComponent[PackType]):
         """
         pass
 
-    def ref_pack_record(self, record_meta: Dict[str, List[str]]):
+    def ref_pack_record(self, record_meta: Dict[str, Set[str]]):
         r"""Method to add output type record of reference datapack of
         current processor to :attr:`forte.data.base_pack.BaseMeta.record`.
